backend/requirements.txt
fastapi==0.115.2
uvicorn[standard]==0.32.0
sqlalchemy==2.0.35
alembic==1.13.2
pydantic==2.9.2
python-dotenv==1.0.1
openpyxl==3.1.5
python-jose[cryptography]==3.3.0
passlib==1.7.4